Refactor: Separate input comment component
- 컴포넌트 재사용성을 위해 input comment component 분리
